ai-service/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.4
easydict==1.13
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.135.2
flatbuffers==25.12.19
fonttools==4.62.1
h11==0.16.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.3
insightface==0.7.3
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnx==1.20.1
onnxruntime==1.23.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
pillow==12.1.1
prettytable==3.17.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.33.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
simsimd==6.5.16
six==1.17.0
starlette==1.0.0
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
azure-storage-blob